% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{data}
\name{comm_palettes}
\alias{comm_palettes}
\title{Comm associations palette}
\format{
An object of class \code{list} of length 2.
}
\usage{
comm_palettes
}
\description{
Comm associations palette
}
\keyword{datasets}
