% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{get_palette_comm}
\alias{get_palette_comm}
\title{Get palette function}
\usage{
get_palette_comm(palette = "ica", reverse = FALSE, select = NULL)
}
\arguments{
\item{palette}{Palette name as a string. See the palettes section.}

\item{reverse}{Should the palette order be reversed?}

\item{select}{An optional string vector of color names that indicates
a custom selection of available colors and their order. See
\url{https://github.com/joon-e/allespaletti}
for all color names.}
}
\description{
Get palette function
}
\keyword{internal}
