% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{get_palette_tol}
\alias{get_palette_tol}
\title{Get palette function}
\usage{
get_palette_tol(
  palette = "vibrant",
  reverse = FALSE,
  select = NULL,
  discrete = TRUE,
  ...
)
}
\arguments{
\item{palette}{Palette name as a string. See the palettes section.}

\item{reverse}{Should the palette order be reversed?}

\item{select}{An optional string vector of color names that indicates
a custom selection of available colors and their order. See
\url{https://github.com/joon-e/allespaletti}
for all color names.}

\item{discrete}{Set to \code{FALSE} if plotting continuous data.}

\item{...}{Other arguments passed on scale function to control name,
limits, breaks, labels and so forth.}
}
\description{
Get palette function
}
\keyword{internal}
