% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politics.R
\docType{data}
\name{politics_palettes}
\alias{politics_palettes}
\title{Politics palette}
\format{
An object of class \code{list} of length 6.
}
\usage{
politics_palettes
}
\description{
Politics palette
}
\keyword{datasets}
