% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{scale_colour_comm}
\alias{scale_colour_comm}
\alias{scale_color_comm}
\alias{scale_fill_comm}
\title{Palettes of communication associations}
\usage{
scale_colour_comm(palette = "ica", reverse = FALSE, select = NULL, ...)

scale_color_comm(palette = "ica", reverse = FALSE, select = NULL, ...)

scale_fill_comm(palette = "ica", reverse = FALSE, select = NULL, ...)
}
\arguments{
\item{palette}{Palette name as a string. See the palettes section.}

\item{reverse}{Should the palette order be reversed?}

\item{select}{An optional string vector of color names that indicates
a custom selection of available colors and their order. See
\url{https://github.com/joon-e/allespaletti}
for all color names.}

\item{...}{Other arguments passed on scale function to control name,
limits, breaks, labels and so forth.}
}
\description{
Color palettes for various communication associations
}
\section{Palettes}{


The following palettes are available for use with these scales:
\describe{
\item{Communication Associations}{ica, dgpuk}
}
}

