% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politics.R
\name{scale_colour_politics}
\alias{scale_colour_politics}
\alias{scale_color_politics}
\alias{scale_fill_politics}
\title{Palettes of political systems}
\usage{
scale_colour_politics(palette = "german", reverse = FALSE, select = NULL, ...)

scale_color_politics(palette = "german", reverse = FALSE, select = NULL, ...)

scale_fill_politics(palette = "german", reverse = FALSE, select = NULL, ...)
}
\arguments{
\item{palette}{Palette name as a string. See the palettes section.}

\item{reverse}{Should the palette order be reversed?}

\item{select}{An optional string vector of color names that indicates
a custom selection of available colors and their order. See
\url{https://github.com/joon-e/allespaletti}
for all color names.}

\item{...}{Other arguments passed on scale function to control name,
limits, breaks, labels and so forth.}
}
\description{
Color palettes for political parties in various countries.
}
\section{Palettes}{


The following palettes are available for use with these scales:
\describe{
\item{German parties}{german, german2, fckafd, fckafd2}
\item{US parties}{us, us2}
}
}

