% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{scale_colour_tol}
\alias{scale_colour_tol}
\alias{scale_color_tol}
\alias{scale_fill_tol}
\title{Sequential, diverging and qualitative colour scales from Paul Tol}
\usage{
scale_colour_tol(
  palette = "vibrant",
  reverse = FALSE,
  select = NULL,
  discrete = TRUE,
  ...
)

scale_color_tol(
  palette = "vibrant",
  reverse = FALSE,
  select = NULL,
  discrete = TRUE,
  ...
)

scale_fill_tol(
  palette = "vibrant",
  reverse = FALSE,
  select = NULL,
  discrete = TRUE,
  ...
)
}
\arguments{
\item{palette}{Palette name as a string. See the palettes section.}

\item{reverse}{Should the palette order be reversed?}

\item{select}{An optional string vector of color names that indicates
a custom selection of available colors and their order. See
\url{https://github.com/joon-e/allespaletti}
for all color names.}

\item{discrete}{Set to \code{FALSE} if plotting continuous data.}

\item{...}{Other arguments passed on scale function to control name,
limits, breaks, labels and so forth.}
}
\description{
Qualitative, sequential, and divergent color schemes that are both
beautiful and distinct for most people (e.g., various forms of
color-blindness) and situations (screen, print-out, black and white copy).
}
\section{Palettes}{


The following palettes are available for use with these scales:
\describe{
\item{Qualitative}{bright, high_contrast, vibrant, muted, pale, dark, light}
\item{Diverging}{sunset, BuRd, PRGn}
\item{Sequential}{YlOrBr, iridescent, rainbow14, rainbow23, smooth_rainbow}}
}

\references{
\url{https://personal.sron.nl/~pault/}
}
